# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Flamoral Dating Platform API
  description: |
    Complete API documentation for the Flamoral Dating Platform.

    This dating platform features:
    - User authentication and profiles
    - Discovery and matching system
    - Real-time messaging
    - Subscription and coins system
    - Boost products
    - Privacy controls
    - Notifications
    - Payment processing
  version: 1.0.0
  contact:
    name: Flamoral Platform Team
    email: platform@flamoral.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.flamoral.com
    description: Production API
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: User login
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: User logout
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/refresh-token:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
  /api/v1/auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email address
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Email verified successfully
  /api/v1/auth/resend-verification:
    post:
      tags: [Auth]
      summary: Resend verification email
      operationId: resendVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Verification email sent
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Password reset email sent
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password reset successful
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/matches:
    get:
      tags: [Matches]
      summary: Get all matches
      operationId: getMatches
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of matches
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMatches'
...
  /api/v1/matches/count:
    get:
      tags: [Matches]
      summary: Get match count
      operationId: getMatchCount
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Match count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
  /api/v1/matches/{matchId}:
    get:
      tags: [Matches]
      summary: Get match details
      operationId: getMatchDetails
      security:
        - bearerAuth: []
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Match details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
    delete:
      tags: [Matches]
      summary: Unmatch user
      operationId: unmatchUser
      security:
        - bearerAuth: []
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Unmatched successfully
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
    RegisterRequest:
      type: object
      required: [email, password, firstName, dateOfBirth, gender]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        accessToken:
          type: string
        refreshToken:
          type: string
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Photo:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        isPrimary:
          type: boolean
        order:
          type: integer
...
    DiscoveryProfile:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        age:
          type: integer
        bio:
          type: string
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        distance:
          type: number
...
    Match:
      type: object
      properties:
        id:
          type: string
        user:
          $ref: '#/components/schemas/DiscoveryProfile'
        matchedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        conversationId:
          type: string
    PaginatedMatches:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Match'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer